# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CTF Full V1 API
  version: 0.1.0
  description: >
    Partial contract skeleton for VS01-VS05 and the read-only External Reality
    Interface. All paths below are relative to the /api/v1 server URL. The runnable
    FastAPI application's generated /openapi.json is authoritative where this
    skeleton does not yet declare an implemented endpoint or exact response shape.
    Domain implementations must expand schemas without weakening ownership, version,
    evidence, genealogy, idempotency, or human-authority requirements. Every
    consequential mutation requires Idempotency-Key in the deployed policy and
    may return 413 or 429 with a stable error envelope.
...
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /projects:
    post:
      tags: [Projects]
      operationId: createProject
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateProject"}
      responses:
        "201": {$ref: "#/components/responses/ProjectResponse"}
        "409": {$ref: "#/components/responses/Conflict"}
  /projects/{projectId}:
    parameters: [{$ref: "#/components/parameters/ProjectId"}]
    get:
      tags: [Projects]
      operationId: getProject
      responses:
        "200": {$ref: "#/components/responses/ProjectResponse"}
        "404": {$ref: "#/components/responses/NotFound"}
  /projects/{projectId}/workspace:
    parameters: [{$ref: "#/components/parameters/ProjectId"}]
    get:
      tags: [Projects]
      operationId: getWorkspace
      responses:
        "200":
          description: Current server-authoritative workspace
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Workspace"}
  /projects/{projectId}/input:
    parameters: [{$ref: "#/components/parameters/ProjectId"}]
    post:
      tags: [Projects]
      operationId: submitProjectInput
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}, {$ref: "#/components/parameters/ExpectedVersion"}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: {type: string, minLength: 1, maxLength: 20000}
      responses:
        "202": {$ref: "#/components/responses/WorkspaceResponse"}
        "409": {$ref: "#/components/responses/Conflict"}
  /projects/{projectId}/gates/{gateId}/decision:
    parameters:
      - {$ref: "#/components/parameters/ProjectId"}
      - {name: gateId, in: path, required: true, schema: {type: string}}
    post:
      tags: [Gates]
      operationId: decideHumanGate
      description: >
        Atomically verifies ownership, active pending gate, allowed choice and exact
        object versions; records the human decision and transition. AI/service
        principals may not call this operation.
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}, {$ref: "#/components/parameters/ExpectedVersion"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/GateDecisionInput"}
      responses:
        "200":
          description: Immutable human decision and resulting workspace
          content:
            application/json:
              schema:
                type: object
                required: [decision, workspace]
                properties:
                  decision: {$ref: "#/components/schemas/GateDecision"}
                  workspace: {$ref: "#/components/schemas/Workspace"}
        "409": {$ref: "#/components/responses/Conflict"}
  /projects/{projectId}/attachments:
    parameters: [{$ref: "#/components/parameters/ProjectId"}]
    get:
      tags: [Attachments]
      operationId: listAttachments
      responses:
        "200":
          description: Project-owned attachment metadata
          content:
            application/json:
              schema: {type: array, items: {$ref: "#/components/schemas/Attachment"}}
    post:
      tags: [Attachments]
      operationId: uploadAttachment
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, document_type]
              properties:
                file: {type: string, format: binary}
                document_type: {$ref: "#/components/schemas/DocumentType"}
      responses:
        "201":
          description: Quarantined/stored attachment metadata; not semantic evidence
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Attachment"}
        "413": {$ref: "#/components/responses/ErrorResponse"}
...
  /projects/{projectId}/attachments/{attachmentId}/download:
    parameters:
      - {$ref: "#/components/parameters/ProjectId"}
      - {name: attachmentId, in: path, required: true, schema: {type: string}}
    get:
      tags: [Attachments]
      operationId: downloadAuthorizedAttachment
      description: Requires project ownership and a clean malware scan. Streams local content or returns an audited short-lived S3 presigned URL.
      responses:
        "200": {description: Binary stream or presigned delivery contract}
        "403": {$ref: "#/components/responses/ErrorResponse"}
        "409": {$ref: "#/components/responses/Conflict"}
...
components:
  securitySchemes:
    bearerAuth: {type: apiKey, in: header, name: X-Session-Token}
    providerCredential: {type: apiKey, in: header, name: X-Provider-Credential}
  parameters:
    ProjectId: {name: projectId, in: path, required: true, schema: {type: string, format: uuid}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 1, maxLength: 200}}
    ExpectedVersion: {name: If-Match, in: header, required: true, schema: {type: string}, description: Expected aggregate version}
  responses:
    ProjectResponse:
      description: Project
      content: {application/json: {schema: {$ref: "#/components/schemas/Project"}}}
    WorkspaceResponse:
      description: Workspace
      content: {application/json: {schema: {$ref: "#/components/schemas/Workspace"}}}
    NotFound: {$ref: "#/components/responses/ErrorResponse"}
    Conflict: {$ref: "#/components/responses/ErrorResponse"}
...
    ErrorResponse:
      description: Stable safe error
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
  schemas:
...
    CreateProject:
      type: object
      required: [entry_family, initial_input]
      properties:
        entry_family: {enum: [CREATION, FUNDING, DOCUMENT]}
        entry_type: {type: string}
        initial_input: {type: string, minLength: 1}
        source: {type: object, additionalProperties: {type: string}}
    Project:
      type: object
      required: [id, stage, version, created_at]
      properties:
        id: {type: string, format: uuid}
        stage: {type: string}
        version: {type: integer, minimum: 1}
        created_at: {type: string, format: date-time}
    Workspace:
      type: object
      required: [project, active_gate, current_cycle]
      properties:
        project: {$ref: "#/components/schemas/Project"}
        active_gate: {oneOf: [{$ref: "#/components/schemas/Gate"}, {type: "null"}]}
        current_cycle: {type: [object, "null"]}
        memory_view: {type: object}
    Gate:
      type: object
      required: [id, number, status, allowed_decisions, subject_versions]
      properties:
        id: {type: string}
        number: {type: integer, minimum: 1, maximum: 19}
        status: {enum: [PENDING, CONFIRMED, SUPERSEDED]}
        allowed_decisions: {type: array, items: {type: string}, minItems: 1}
        subject_versions: {type: object, additionalProperties: {type: integer}}
    GateDecisionInput:
      type: object
      required: [decision, subject_versions]
      properties:
        decision: {type: string}
        rationale: {type: string, maxLength: 10000}
        subject_versions: {type: object, additionalProperties: {type: integer}}
    GateDecision:
      type: object
      required: [id, gate_id, decision, origin, decided_at]
      properties:
        id: {type: string}
        gate_id: {type: string}
        decision: {type: string}
        origin: {const: USER}
        rationale: {type: string}
        decided_at: {type: string, format: date-time}
    DocumentType:
      enum: [CALL_DOCUMENT, APPLICATION, PARTIAL_APPLICATION, PROJECT_CONCEPT, TECHNICAL_DOCUMENT, FINANCIAL_DOCUMENT, OTHER]
    Attachment:
      type: object
      required: [id, project_id, original_filename, mime_type, size, checksum, status]
      properties:
        id: {type: string}
        project_id: {type: string, format: uuid}
        original_filename: {type: string}
        mime_type: {type: string}
        size: {type: integer, minimum: 0}
        checksum: {type: string}
        status: {enum: [SCANNING, READY_FOR_LATER_PROCESSING, ANALYZED, QUARANTINED, DELETED]}
        malware_scan: {type: object}
        semantic_analysis_claimed: {type: boolean, const: false}
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: {type: string}
            message: {type: string}
            request_id: {type: string}